+++ query-connector/src/app/layout.tsx
@@ -1,8 +1,8 @@
 import "../styles/styles.scss";
 import Header from "./query/components/header/header";
 import Footer from "./query/components/footer/footer";
-import { DataProvider } from "./utils";
 import { SessionProvider } from "next-auth/react";
+import DataProvider from "./DataProvider";
 
 /**
  * Establishes the layout for the application.
